+++ ffun/ffun/user_settings/migrations/20240909_01_tQXMR-replace-tokens-traking-with-costs-tracking.py
@@ -20,7 +20,7 @@ cost_1m = 0.6
 
 
 def tokens_to_costs(tokens: int) -> int:
-    return tokens / 1_000_000 * cost_1m
+    return int(tokens / 1_000_000 * cost_1m)
 
 
 def apply_step(conn: Connection[dict[str, Any]]) -> None:
